third_party/blink/web_tests/fast/canvas/color-space/canvas-getImageData.html

<!DOCTYPE html>
<body>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
test(drawThenGetImageData,
    'tests drawing SRGB color and reading it back using getImage.');

function run_test(canvas_color_space, canvas_pixel_format, data_color_space, data_storage_format) {
  var color_style = 'rgb(50, 100, 150)';

  var color_srgb_f32 = [50/255., 100/255., 150/255., 1.0];
  var color_srgb_u8  = [color_srgb_f32[0] * 255, color_srgb_f32[1] * 255, color_srgb_f32[2] * 255, 255];
  var color_srgb_u16 = [color_srgb_f32[0] * 65535, color_srgb_f32[1] * 65535, color_srgb_f32[2] * 65535, 65535];

  var color_p3_f32 = [0.24304, 0.38818, 0.57227, 1.0];
  var color_p3_u8  = [color_p3_f32[0] * 255, color_p3_f32[1] * 255, color_p3_f32[2] * 255, 255];
  var color_p3_u16 = [color_p3_f32[0] * 65535, color_p3_f32[1] * 65535, color_p3_f32[2] * 65535, 65535];

  var color_rec2020_f32 = [0.34106, 0.41553, 0.59180, 1.0];
  var color_rec2020_u8  = [color_rec2020_f32[0] * 255, color_rec2020_f32[1] * 255, color_rec2020_f32[2] * 255, 255];
  var color_rec2020_u16 = [color_rec2020_f32[0] * 65535, color_rec2020_f32[1] * 65535, color_rec2020_f32[2] * 65535, 65535];

  var epsilon_f32 =  2 / 255.;
  var epsilon_u8 = 2;
  var epsilon_u16 = 2 * 255;

  var canvas = document.createElement('canvas');
  canvas.width = 10;
  canvas.height = 10;
  var ctx = canvas.getContext('2d', {colorSpace: canvas_color_space, pixelFormat:canvas_pixel_format});
  ctx.fillStyle = color_style;
  ctx.fillRect(0, 0, 10, 10);

  var pixel = ctx.getImageData(5, 5, 1, 1, {colorSpace: data_color_space, storageFormat:data_storage_format}).data;

  var pixel_expected;
  var epsilon;
  if (data_storage_format == 'float32') {
    epsilon = epsilon_f32;
    if (data_color_space == 'srgb')
        pixel_expected = color_srgb_f32;
    if (data_color_space == 'p3')
        pixel_expected = color_p3_f32;
    if (data_color_space == 'rec2020')
        pixel_expected = color_rec2020_f32;
  } else if (data_storage_format == 'uint16') {
    epsilon = epsilon_u16;
    if (data_color_space == 'srgb')
        pixel_expected = color_srgb_u16;
    if (data_color_space == 'p3')
        pixel_expected = color_p3_u16;
    if (data_color_space == 'rec2020')
        pixel_expected = color_rec2020_u16;
  } else if (data_storage_format == 'uint8') {
    epsilon = epsilon_u8;
    if (data_color_space == 'srgb')
        pixel_expected = color_srgb_u8;
    if (data_color_space == 'p3')
        pixel_expected = color_p3_u8;
    if (data_color_space == 'rec2020')
        pixel_expected = color_rec2020_u8;
  }

  var passed = true;
  assert_true(pixel.length === pixel_expected.length);
  for (var i = 0; i < pixel.length; i++)
    if (Math.abs(pixel[i] - pixel_expected[i]) > epsilon)
      passed = false;
  if (!passed) {
    console.log('Failure!');
    console.log('  Canvas:    ' + canvas_color_space + ',' + canvas_pixel_format);
    console.log('  ImageData: ' + data_color_space + ',' + data_storage_format);
    console.log('  Expected:  ' + pixel_expected);
    console.log('  Actual:    ' + [pixel[0], pixel[1], pixel[2], pixel[3]]);
    console.log('  Epsilon:   ' + epsilon);
  }
  return passed;
}

function drawThenGetImageData() {
  color_spaces = ['srgb', 'p3', 'rec2020'];
  canvas_pixel_formats = ['uint8', 'float16'];
  image_data_pixel_formats = ['uint8', 'uint16', 'float32'];
  var passed = true;
  for (var i = 0; i < 3; ++i) {
    for (var j = 0; j < 2; ++j) {
      for (var k = 0; k < 3; ++k) {
        for (var l = 0; l < 3; ++l) {
          // TODO(http://crbug.com/skia/10965): Reading a uint8 canvas from a
          // uint8 canvas with no color conversion will cause an unexpected
          // transformation to be performed on the data.
          if (i == k && canvas_pixel_formats[j] == 'uint8' && image_data_pixel_formats[l] == 'uint8')
            continue;
          if (!run_test(color_spaces[i], canvas_pixel_formats[j], color_spaces[k], image_data_pixel_formats[l]))
            passed = false;
        }
      }
    }
  }
  assert_true(passed);
}
</script>
</body>
